requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.0.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
spotipy>=2.23.0
aiohttp>=3.9.0
